+++ CHANGELOG.md
@@ -7,6 +7,9 @@
   ([#186](https://github.com/biocore/qurro/issues/186))
 - Added feature rankings as searchable fields in the "Selecting Features"
   controls.
+- Feature metadata fields and feature ranking fields are now grouped under
+  `Feature Metadata` or `Feature Ranking` headers in the searchable fields
+  dropdowns. ([#191](https://github.com/biocore/qurro/issues/191))
 - Added numeric searching: now you can search through numeric feature metadata or
   feature rankings using basic comparison operators. (Non-numeric input search
   text will result in the search not identifying any features, and non-numeric